add /[username]/p/[postUrlName] route for post page, redirect old url
